Deeper scan part 2: Convert pandas/numpy values to float in sector_analyzer, macro_analyzer, multi_asset_analyzer, and data_layer macro data

+++ data_layer.py
@@ -368,8 +368,8 @@ class DataLayer:
                 hist = ticker.history(period="1mo")
                 if not hist.empty:
                     data[name] = {
-                        'current': hist['Close'].iloc[-1],
-                        'change_1m': (hist['Close'].iloc[-1] / hist['Close'].iloc[0] - 1) * 100
+                        'current': float(hist['Close'].iloc[-1]),
+                        'change_1m': float((hist['Close'].iloc[-1] / hist['Close'].iloc[0] - 1) * 100)
                     }
             except:
                 continue
